Make GlyphManager switchable.
